Extract LoopRenderer
